# internal/transport/rest/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/prices/storage": {
            "get": {
                "description": "Returns normalized storage pricing across cloud providers (AWS, Azure, GCP) for requested specs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prices"
                ],
                "summary": "Get storage pricing comparison",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Requested storage size in GB (e.g. 500, max 1000000)",
                        "name": "size_gb",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Requested canonical storage class (e.g. standard, infrequent_access, archive)",
                        "name": "storage_class",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Canonical region group (default: us-east)",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Target currency code (default: USD)",
                        "name": "currency",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.StorageComparisonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.RFC7807Error"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/middleware.RFC7807Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.RFC7807Error"
                        }
                    }
                }
            }
        },
        "/api/v1/providers/{provider}/status": {
            "get": {
                "description": "Returns the latest successful fetch timestamp, per-category observation count, staleness flag, and active DLQ failure records for the specified cloud provider.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "providers"
                ],
                "summary": "Get cloud provider operational status and data freshness",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud provider identifier (e.g. aws, azure, gcp, oracle, ibm, alibaba, digitalocean)",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Provider status and category freshness details",
                        "schema": {
                            "$ref": "#/definitions/rest.ProviderStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Parameters",
                        "schema": {
                            "$ref": "#/definitions/middleware.RFC7807Error"
                        }
                    },
                    "404": {
                        "description": "Provider Not Found",
                        "schema": {
                            "$ref": "#/definitions/middleware.RFC7807Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.RFC7807Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.StorageAttributes": {
            "type": "object",
            "properties": {
                "iops": {
                    "description": "IOPS holds provisioned IOPS where reported (e.g. block storage EBS gp3/io2).\nCurrently unpopulated for S3/Blob/GCS object storage in stage 1.4.",
                    "type": "integer"
                },
                "size_gb": {
                    "type": "number"
                },
                "storage_class": {
                    "type": "string"
                }
            }
        },
        "middleware.RFC7807Error": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "rest.CategoryStatusResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "compute"
                },
                "dlq": {
                    "$ref": "#/definitions/rest.DLQStatusResponse"
                },
                "last_fetched_at": {
                    "type": "string"
                },
                "last_seen_at": {
                    "type": "string"
                },
                "observation_count": {
                    "type": "integer",
                    "example": 450
                },
                "stale": {
                    "type": "boolean",
                    "example": false
                },
                "staleness_threshold_hours": {
                    "type": "number",
                    "example": 168
                },
                "supported": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "rest.DLQStatusResponse": {
            "type": "object",
            "properties": {
                "consecutive_failures": {
                    "type": "integer",
                    "example": 3
                },
                "last_error": {
                    "type": "string",
                    "example": "503 Service Unavailable"
                },
                "status": {
                    "type": "string",
                    "example": "failed"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "rest.PriceDetail": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "rest.ProviderStatusResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/rest.CategoryStatusResponse"
                    }
                },
                "last_successful_fetch": {
                    "type": "string"
                },
                "provider": {
                    "type": "string",
                    "example": "aws"
                },
                "stale": {
                    "type": "boolean",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "example": "healthy"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.ProviderStatusWarningResponse"
                    }
                }
            }
        },
        "rest.ProviderStatusWarningResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "not_yet_ingested"
                },
                "message": {
                    "type": "string",
                    "example": "Oracle OCI ingestion lands in stage 4."
                },
                "provider": {
                    "type": "string",
                    "example": "oracle"
                }
            }
        },
        "rest.ProviderWarning": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
...
        "rest.StorageComparisonMeta": {
            "type": "object",
            "properties": {
                "api_version": {
                    "type": "string"
                },
                "generated_at": {
                    "type": "string"
                },
                "query": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "rest.StorageComparisonResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/rest.StorageComparisonMeta"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.StorageResultEntry"
                    }
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.ProviderWarning"
                    }
                }
            }
        },
        "rest.StorageResultEntry": {
            "type": "object",
            "properties": {
                "fetched_at": {
                    "type": "string"
                },
                "match_delta_pct": {
                    "type": "number"
                },
                "match_quality": {
                    "type": "string"
                },
                "matched_spec": {
                    "$ref": "#/definitions/domain.StorageAttributes"
                },
                "missing_attributes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "monthly_cost_usd": {
                    "type": "number"
                },
                "price": {
                    "$ref": "#/definitions/rest.PriceDetail"
                },
                "provider": {
                    "type": "string"
                },
                "sku_id": {
                    "type": "string"
                },
                "stale": {
                    "type": "boolean"
                }
            }
        }
    }
}